Do not overwrite package manager config files if they exist
Let's allow users to override mkosi's default package manager
settings by providing their own package manager config file. This
is especially important for distributions built using dnf as dnf
does not yet have support for dropins.

+++ mkosi/distributions/arch.py
@@ -63,6 +63,9 @@ def setup_pacman(state: MkosiState) -> None:
     state.root.joinpath("var/lib/pacman").mkdir(mode=0o755, exist_ok=True, parents=True)
 
     config = state.pkgmngr / "etc/pacman.conf"
+    if config.exists():
+        return
+
     config.parent.mkdir(mode=0o755, exist_ok=True, parents=True)
 
     with config.open("w") as f:
